refactor(server): 第四阶段：抽取多处数据库操作逻辑至服务层并优化命令实现
- WhoDropsCommand改用DropDataService获取掉落来源数据，移除手动数据库连接代码
- UnBanCommand改用AccountService统一解除封禁逻辑，避免多处数据库操作
- PmobCommand和PnpcCommand新增调用PlifeService处理地图生命体持久化，替代原生SQL
- PmobRemoveCommand和PnpcRemoveCommand改用PlifeService封装的移除逻辑，减少原生SQL代码
- ItemFactory移除大量手写JDBC逻辑，改为调用ItemFactoryService统一加载和保存物品数据
- Pet类移除数据库连接与操作代码，改用PetService提供宠物数据加载、保存和删除功能
- 移除SQL异常抛出，在服务层统一处理异常，调用处捕获打印错误信息
- 去除多个类中的数据库连接、预处理语句和结果集的复杂操作，提升代码可维护性
- 保持原有功能不变，保证向后兼容性及运行逻辑一致性

+++ gms-server/src/main/java/org/gms/dao/entity/EventstatsDO.java
@@ -27,7 +27,7 @@ public class EventstatsDO implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id
-    private Long characterid;
+    private Integer characterid;
 
     /**
      * 0
